Extract core into separate library
All of the core code has been extracted into a separate
`TelemetryMetricsPrometheus.Core` library. This library is backwards compatible
for existing users, starting both the `TelemetryMetricsPrometheus.Core` library
as well as the cowboy server.

The docs have been updated to reference the `TelemetryMetricsPrometheus.Core`
docs where appropriate.

The tests have been updated to stop both `TelemetryMetricsPrometheus`
and `TelemetryMetricsPrometheus.Core`. The stop is now called on an
`on_exit` to ensure that a test failure does not impact stopping the
reporter.